refactor: update code in accordance with mentor's comments
