+++ src/App.tsx
@@ -1,15 +1,17 @@
 import { Route, BrowserRouter, Routes } from 'react-router-dom';
-import HomePage from './pages/HomePage';
-import Details from './react-components/Details/Details';
+import HomePage from './pages/HomePage/HomePage';
+import Details from './pages/Details/Details';
 
 export default function App() {
   return (
     <BrowserRouter>
-      <Routes>
-        <Route path="/" element={<HomePage />}>
-          <Route path="details/:id" element={<Details />} />
-        </Route>
-      </Routes>
+      <div className="app">
+        <Routes>
+          <Route path="/" element={<HomePage />}>
+            <Route path="details/:id" element={<Details />} />
+          </Route>
+        </Routes>
+      </div>
     </BrowserRouter>
   );
 }
