+++ src/pages/Details/Details.css
