+++ src/pages/Details/types.ts
