+++ src/pages/HomePage/HomePage.css
@@ -1,0 +1,4 @@
+.app {
+  height: 100vh;
+  background-color: rgb(160, 216, 160);
+}
