+++ src/react-components/Header/Header.css
@@ -1,5 +1,5 @@
 .search-line {
-  padding-bottom: 30px;
+  padding-bottom: 5vh;
   padding-top: 5vh;
   width: 100%;
   display: flex;
