+++ src/react-components/SearchResults/SearchResults.css
@@ -38,6 +38,8 @@
   justify-content: center;
   align-items: center;
   gap: 20px;
+  text-decoration: none;
+  color: black;
 }
 
 .search-results.with-details {
@@ -54,6 +56,7 @@
   color: black;
 }
 
+.search-results:visited,
 .product-container:visited {
   text-decoration: none;
   color: black;
